requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.5.0
flask>=2.3.0
requests>=2.28.0
tqdm>=4.65.0
